+++ hexrd/fitting/calibration/__init__.py
@@ -1,0 +1,16 @@
+from .instrument import InstrumentCalibrator
+from .multigrain import calibrate_instrument_from_sx, generate_parameter_names
+from .powder import PowderCalibrator
+from .structureless import StructurelessCalibrator
+
+# For backward-compatibility, since it used to be named this:
+StructureLessCalibrator = StructurelessCalibrator
+
+__all__ = [
+    'calibrate_instrument_from_sx',
+    'generate_parameter_names',
+    'InstrumentCalibrator',
+    'PowderCalibrator',
+    'StructurelessCalibrator',
+    'StructureLessCalibrator',
+]
